Initial commit
This commit kicks off the migration from Middleman to Hugo for my
personal site: lecklider.com. Since this is my first time working with
Hugo, I'm sure this stuff will change, but this is a decent start.

Some of the layout files are based off of the Type theme
https://github.com/digitalcraftsman/hugo-type-theme. Many thanks to the
authors of that theme for paving the way for me.